feat: 상품의 position이 아닌 id를 비교하여 상품을 삭제하도록 구현

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -17,7 +17,7 @@ class DefaultCartRepository(
         return dao.getAll().map { it.toDomain() }
     }
 
-    override suspend fun deleteCartProduct(id: Int) {
-        dao.delete(id.toLong())
+    override suspend fun deleteCartProduct(id: Long) {
+        dao.delete(id)
     }
 }
