+++ domain/src/main/java/woowacourse/shopping/repository/CartRepository.kt
@@ -6,5 +6,5 @@ import woowacourse.shopping.model.Product
 interface CartRepository {
     suspend fun addCartProduct(product: Product)
     suspend fun getAllCartProducts(): List<CartProduct>
-    suspend fun deleteCartProduct(id: Int)
+    suspend fun deleteCartProduct(id: Long)
 }
